4-19-18 All swing values finalized as doubles

+++ java/golf/golf_app_2/Coordinates.kt
@@ -1,9 +1,9 @@
 package golf.golf_app_2
 
 class Coordinates {
-    var cood_x: Int = 0
-    var cood_y: Int = 0
-    constructor(cood_X:Int, cood_Y:Int) {
+    var cood_x: Double = 0.0
+    var cood_y: Double = 0.0
+    constructor(cood_X:Double, cood_Y:Double) {
         this.cood_x = cood_X
         this.cood_y = cood_Y
     }
